src/templates/bot_template/requirements.txt
# TeleTask Bot Dependencies
# Generated by install.sh

# Telegram Bot Framework
python-telegram-bot>=21.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Migrations
alembic>=1.13.0

# Scheduler
APScheduler>=3.10.0

# Environment
python-dotenv>=1.0.0

# HTTP Client
aiohttp>=3.9.0

# Timezone
pytz>=2024.1

# Cache (optional)
redis>=5.0.0

# Monitoring (optional)
prometheus-client>=0.19.0
